Python 3.11 + Pygame-ce 2.3.2

See merge request rainlash/lt-maker!991

+++ app/editor/main_editor.py
@@ -91,7 +91,7 @@ class MainEditor(QMainWindow):
         main_screen_size = desktop.availableGeometry(desktop.primaryScreen())
 
         # Use setFixedSize to make it permanent and unchangeable
-        default_size = main_screen_size.width()*0.7, main_screen_size.height()*0.7
+        default_size = int(main_screen_size.width()*0.7), int(main_screen_size.height()*0.7)
         self.resize(*default_size)
 
         geometry = self.settings.component_controller.get_geometry(
